Docs and folders reorganization
