+++ src/collection/fusion.py
